OProfile testsuite, fix white spaces
This patch cleans up trailing white spaces, spaces before tabs.  In the
lib/op_events.exp file, the tables of events were cleaned up to use
tabs and spaces consistently, fix the consistency on the location of the
line contiuation charcter.  Additionally, the patch makes sure that the
files all end with a blank line consisting of a single space.

Signed-off-by: Carl Love <cel@us.ibm.com>
Signed-off-by: Maynard Johnson <maynardj@us.ibm.com>