OProfile testsuite, add operf support to the testsuite
This patch adds support for the basic testing of the new
operf user interface.  It leverages the existing event
tables for the various processors.  OProfile has its own
kernel support for the names of the processors.  This
information is only available when the OProfile kernel
module is loaded.  The operf support gets the CPU name from
the ophelp command instead.

The patch also fixes an error in the existing opcontol testsuite.
Specifically, the test for the string "opcontrol error:" will always
fail because the output from the command does not include the word
"opcontrol".  Removed "opcontrol" from the test string.

Added commands to the opcontrol testsuite to make sure the
daemonrc file is removed.  This is necessary to ensure there are no
extraneous settings, for example filter settings, that will cause errors.

Trailing blanks on five lines in the existing files were removed.

The testsuit for testing OProfile using the opcontrol interface
has been renamed to oprofile-opcontrol.  The new testsuite to
test OProfile using operf is called oprofile-operf.  To run both
testsuites, use the testsuite oprofile.  Specifically the commands
to run the testsuites are:

    runtest --tool oprofile-opcontrol     // test opcontrol interface

    runtest --tool oprofile-operf         // test operf interface

    runtest --tool oprofile               // test both interfaces

Signed-off-by: Carl Love <cel@us.ibm.com>