OProfile testsuite, change operf version check fail to warning
This patch changes the operf version test failure to a warning instead
of a failure.  Also, the user message that is printed is changed to
be more explicit that the user tool operf is not supported rather then
using the term "system" which could be interpreted as meaning the
kernel.  The patch specifically addresses systems with a perf enabled
kernel but an older (OProfile 0.9.7 or older) user space tool that
does not have support for the new operf user interface.

Signed-off-by: Carl Love <cel@us.ibm.com>