OProfile testsuite, fix perf kill command
The operf user tool expects to get a kill -SIGINT signal.  However
the SIGINT signal can not be sent from a script, it must come from
the command line.  Hence the pkill -P command was used to kill the
parent and children processes.  A bug was found in that the pkill
command kills operf without letting it completely finish the data
processing and exit cleanly.  This issue actually was found when
testing a patch that is currently under development to check for
specific symbols.  The kill -INT signal can be sent from the script
and is properly caught by operf.  When operf catches the INT signal
it shuts down cleanly rather then exiting immediatly.  This patch
fixes the current kill command to make sure operf exits cleanly.

Signed-off-by: Carl Love <cel@us.ibm.com>