Oprofile testsuite, fix proc op_daemon_delete_any_samples, it doesn't work correctly
The op_daemon_delete_any_samples procedure is not correctly deleting
samples.  This patch replaces the body of the procedure with the
"opcontrol --reset" call to remove the existing samples.  The patch
also moves the function to lib/op_util.exp so it can be called by
the operf testsuite as well.  A call to op_daemon_delete_any_samples was
added to the operf test main routine to make sure any opcontrol samples are
deleted so opreport will not pickup any samples from a previous opcontrol
run.

Signed-off-by: Carl Love <cel@us.ibm.com>