OProfile testsuite, Add symbol checking functionality
This patch adds additional testing and workload capability. The
patch adds the capability to easily add a new workload specification
and a list of symbols that should be present in the OProfile report.
The testsuite runs each of the specified workloads with each of the
event lists currently specified.  The output from opreport is
checked to verify the workload symbols are found.

Signed-off-by: Carl Love <cel@us.ibm.com>