OProfile testsuite, add single process profiling functionality
This patch adds another test to the suite of tests. The test runs each of
the workloads with operf in single process mode.  The test verifies that
samples were collected for each run.  The test checks to make sure the
symbols were found for the first event group, just as is done in the
oprofile-operf test.  The testsuite can be run as root or as a regular user.

Signed-off-by: Carl Love <cel@us.ibm.com>