Fix tcl workload symbol checking
Intermittently, the operf system-wide and operf single-process
tests fail after running the tcl workload (which executes a dd
command) and attempting to find the string "/bin/dd" in the
opreport output. The reason for this is two-fold:
  - operf will profile child processes that a parent process will
    fork/exec, but will terminate the profiling once the parent
    process exits.
  - The testsuite/workloads/tcl_src/workload.tcl script execs
    the dd command as the last statement in the script and then
    exits.

Considering the two facts stated above, it's clear that profiling
of the dd command may or may not occur, depending on timing.
Once the parent tcl process exits, operf will detect the parent
exiting and will shutdown profiling. Some samples may have been
collected for the dd command before this shutdown occurs, but maybe
not.

To avoid this timing issue, this patch simply changes the symbol
we check for in the tcl profile output from "/bin/dd" to "libtcl".

Signed-off-by: Maynard Johnson <maynardj@us.ibm.com>