Adjust events for IBM POWER7 processor
The POWER7+ processor (which has the same events as POWER7)
has a much larger cache than POWER7.  When running the
testsuite on POWER7+, the PM_BR_PRED_CCACHE event was not
being seen at all.  I removed this event from the list of
events to monitor for since cache-related events are not
easy to deterministically predict for unknown processor
types.  I also adjusted down the number of certain other
POWER7 events to help ensure the testing would "see" some
samples for those events.

Signed-off-by: Maynard Johnson <maynardj@us.ibm.com>