oprofile-tests: Fix error message "Do not know how to handle version {1.0.0git}"
This patch fixes the message "Do not know how to handle version {1.0.0git}".
The current code requires that each new major version be added to the
case statement for versions greater then 0.5.  Apparently, versions prior
to 0.5 were not supported by the testsuite for some now unkown reason.  This
patch creates a specific case to handle the pre 0.5 versions, the 0.5 version
and all newer versions are handled by the default case.  The case for the
pre 0.5 version of profile will simply print a message saying the version
is not supported.  The case for version 0.5 is unchanged.  The default
will handle all newer versions so this issue doesn't have to be continually
fixed with each major version change.

Signed-off-by: Carl Love <cel@us.ibm.com>