OProfile testsuite: Fix event name check in proc get_event_count
The recent patch to fix OProfile bug 253

"ocount does not display user specified unit mask, kerne and user values"

changed the output format when collecting in user or kernel mode only.
This change broke the testsuite function that extracted the count for a
given symbol because the expected symbol failed to match the symbol with
the user and kernel modes appended.  This patch fixes the issue by spliting
the symbol from the output on colons and then checking the string before the
colon.

Signed-off-by: Carl Love <cel@us.ibm.com>