Make new XML tests run on oprofile 1.0 or higher
The schema validation being done in the newly-added XML tests
fails when run against a version of oprofile that does not have
the recent schema fixes.  This patch checks the oprofile version
and skips the XML tests if the version is 0.9.9 or lower.
Additionally, a change was made so that XML tests would only be
run once.  Also made a few indentation corrections.

Signed-off-by: Maynard Johnson <maynardj@us.ibm.com>