OProfile 1.0 test suite, Fix calls to opcontrol
Operf and ocount make calls to opcontrol to make sure the daemon is shut down,
clear out any existing samples etc.  Starting with version 1.0, the
opcontrol command       no longer exists.  This patch will ensure that operf and
ocount  do not make any calls to opcontrol if the command does  not exist.

The opcontrol test suite currently checks if opcontrol exists and exits if
it doesn't exist.  This test suite needs just to set the opcontrol_installed
variable to 0 in the case that opcontrol isn't installed so the exit routine
will not call opcontrol --deinit.

Signed-off-by: Carl Love <cel@us.ibm.com>