OProfile-testsuite, opcontrol version fail fix
Do not fail opcontrol version check if opcontrol is not installed.  Do not
print module info if opcontrol is not installed.

Moved the warning message about needing to be root when running
opcontrol to a point after we determine that opcontrol should,
in fact, be run.

If we cannot determine the version of operf or ocount for any
reason, we print a message and exit immediately.

Signed-off-by: Carl Love <carll@us.ibm.com>