Oprofile-testsuite, skip mode selection tests if not supported.
Event counts may occur when the system is running in user mode, kernel
mode or hypervisor modes.  Some architectures have settings that allow
the user to select which modes the events should be collected.  The
testsuite contains tests to make sure sum of counting in each of the modes
individually is the same as counting in the modes simultaneously.  This
patch skips these tests for architectures that do not support mode selection

Signed-off-by: Carl Love <carll@us.ibm.com>