Oprofile-testsuite, handle power-savings-mode changes during scaling test
The scaling tests will fail if the power-savings-mode changes while the
test is running.  (idle/slow <-> non-idle/fast).
This adds some logic to check the power state before and after running
the ocount scaling workload, and reruns the workload if the state has
changed before it attempts to process results.

Signed-off-by: Will Schmidt <will_schmidt@vnet.ibm.com>