Correct the test events for AMD generic and family15h processors
The default unit mask for the event DATA_CACHE_MISSES  on AMD family 15h
CPUs is 0x1 instead of 0x0, which is common for other AMD CPUs. When the
"athlon" test event table is used on an AMD family 15h machine, the test
fails. A slightly modified event table for AMD family 15h has been added
in order to fix the issue.

Signed-off-by: Michael Petlan <mpetlan@redhat.com>