Use real path to ophelp when searching for other files
Recent versions of Fedora have a symbolic link from /bin to /usr/bin.
Root has /bin before /usr/bin in PATH, so a "which ophelp" will
indicate that ophelp is located in /bin rather than the correct
/usr/bin.  This prevents the tests from using the proper path for
opreport.xsd.  The additional realpath gets the actual path for the
ophelp, so the correct location is searched for opreport.xsd.