Allow testing of oprofile-0.9.9 on platforms that version doc directories
When testing on RHEL7 machines I noticed that the verification test
using the ophelp.xsd files were not occurring or were not finding the
ophelp.xsd file.  The patch expanded the tests to run with
oprofile-0.9.9 to allow some regression testing and expanded the
search for the ophelp.xsd file to find the file on machine that
version the files in doc.