Add test to check correct handling of remapped anon regions
Java Runtime Environments (JRE) generate code during execution and may
be remapping anonymous regions of memory used to store those
translation. OProfile needs to be able to handle those cases.  This
new tests simulates that JRE behavior, by allocating a large chunk of
memory, placing a function in it, and remapping the region that the
generated function is in.

Signed-off-by: William Cohen <wcohen@redhat.com>