Handle the ppc64be use of function descriptors
Rather than directly using pointers to functions the ppc64be uses
pointers to function descriptors. The descriptor is a data structure
with three fields: the actual address of the function, the GOT (Global
Offset Table), and the environment pointer.  The initialize_anon_func
has been modified to return a function descriptor for ppc64be so the
tests works on environements such as RHEL7 on power.

Signed-off-by: William Cohen <wcohen@redhat.com>