swizzle kptr_restrict logic
swizzle kptr_restrict logic

The kptr_restrict field is currently documented as holding one of
three values:
 # 0 - no restriction.
 # 1 - restricted.
 # 2 - addresses are zero'd out.

The existing tests that check the value of the kptr_restrict field
are obsolete in that they don't handle the restricted mode. After some
review and experimentation, simply removing the check simplifies the scripts
and allows more tests to run successfully.  Thus, remove the checks.

Signed-off-by: Will Schmidt <will_schmidt@vnet.ibm.com>