oprofile-tests increase iterations for anon_remap test
Hi,

 We've noticed the anon_remap.c testcase failing on some of our systems.
From a verbose test run log, we note that the anon_remap_bin workload is
getting just a small number of the overall samples.
Results vary greatly depending on the system size and speed.

> samples  %        app name                 symbol name
> 5031     99.7620  /no-vmlinux              /no-vmlinux
> 5         0.0991  /usr/bin/bash            /usr/bin/bash
> 1         0.0198  /home/willschm/oprofile-tests/testsuite/workloads/anon_remap_src/anon_remap_bin 00000017.plt_call.memcpy@@GLIBC_2.17
> 1         0.0198  /home/willschm/oprofile-tests/testsuite/workloads/anon_remap_src/anon_remap_bin initialize_anon_func

The (anon_remap.c) testcase itself will take an argument to override the
default number of iterations, but as far as I can tell the test harness
does not handle passing arguments to the test itself.
I've instead increased the default iteration to a much higher
value, this seems sufficient to allow the test to pass.

Signed-off-by: Will Schmidt <will_schmidt@vnet.ibm.com>

--

 anon_remap.c |    2 +-
 1 file changed, 1 insertion(+), 1 deletion(-)